Adding pagination

+++ responses/pagination.res.go
@@ -1,7 +1,7 @@
 package responses
 
 type PaginationResponse struct {
-	Limit int `json:"limit"`
-	Page  int `json:"page"`
+	Limit int64 `json:"limit"`
+	Page  int64 `json:"page"`
 	GenericResponse
 }
